% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/point_Hobs.R
\name{point_Hobs}
\alias{point_Hobs}
\title{Calculate observed heterozygosity from maximum \emph{a posteriori} genotype estimates}
\usage{
point_Hobs(genotypes, ploidy)
}
\arguments{
\item{genotypes}{A matrix of estimated genotypes returned from the function \code{\link{get_map_genotypes}}.}

\item{ploidy}{The ploidy level of individuals in the population (must be >= 2).}
}
\value{
Returns per locus estimates of observed heterozygosity (\code{per_locus_Hobs}).
}
\description{
Uses point estimates of genotypes calculated with \code{\link{get_map_genotypes}} to calculate the per locus and overall observed heterozygosity.
}
\references{
Hardy, OJ. 2015. Population genetics of autopolyploids under a mixed mating model and the estimation of selfing rate. \emph{Molecular Ecology Resources}, doi: 10.1111/1755-0998.12431.
}

