% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/polyfreqs.R
\name{polyfreqs}
\alias{polyfreqs}
\title{Bayesian estimation of allele frequencies in polyploids}
\usage{
polyfreqs(tM, rM, ploidy, iter = 50000, thin = 100, error = 0.01,
  genotypes = FALSE, col_header = "", outfile = "polyfreqs-mcmc.out")
}
\arguments{
\item{tM}{Total reads matrix: matrix containing the total number of reads mapping to each locus for each individual.}

\item{rM}{Reference reads marix: matrix containing the number of reference reads mapping to each locus for each individual.}

\item{ploidy}{The ploidy level of the sampled individuals (must be >= 2).}

\item{iter}{The number of MCMC generations to run (default=50,000).}

\item{thin}{Thins the MCMC output by sampling everything \code{thin} generations (default=100).}

\item{error}{The level of sequencing error. A fixed constant (default=0.01).}

\item{genotypes}{Logical variable indicating whether or not to print the values of the genotypes sampled during the MCMC (default=FALSE).}

\item{col_header}{Optional column header tag for use in running loci in parallel (default="").}

\item{outfile}{The name of the ouput file that samples from the posterior distribution of allele frequencies are written to (default="polyfreqs-mcmc.out").}
}
\description{
\code{polyfreqs} implements a Gibbs sampler to perform Bayesian inference on the allele frequencies in a population of autopolyploids.
}
\examples{
data(total_reads)
data(ref_reads)
polyfreqs(total_reads,ref_reads,4,iter=100,thin=10)
}
\author{
Paul Blischak
}

