% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/polyfreqs_pps.R
\name{polyfreqs_pps}
\alias{polyfreqs_pps}
\title{Posterior predictive model checks for the model of inheritance}
\usage{
polyfreqs_pps(p_post, tM, rM, ploidy, error)
}
\arguments{
\item{p_post}{A matrix containing the posterior samples from a \code{polyfreqs} run.}

\item{tM}{Total reads matrix.}

\item{rM}{Reference reads matrix.}

\item{ploidy}{Ploidy level of individuals in the population.}

\item{error}{Probability of a sequencing error.}
}
\description{
Uses the posterior distribution of allele frequences from a \code{polyfreqs} run to test model fit using posterior predictive simulation.
}

