% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/posterior_Hexp.R
\name{posterior_Hexp}
\alias{posterior_Hexp}
\title{Calculate posterior distribution of expected heterozygosity from posterior samples of allele frequencies}
\usage{
posterior_Hexp(p_post)
}
\arguments{
\item{p_post}{A matrix of posterior samples of allele frequencies returned by \code{\link{polyfreqs}}.}
}
\value{
A list with two items: the posterior distribution for the overall expected heterozygosity across loci (\code{Hexp_post}) and a matrix with the per locus posterior samples of expected heterozygosity (\code{per_locus_Hexp_post}).
}
\description{
Estimates a posterior distribution for the expected heterozygosity using the poterior samples of allele frequencies returned by \code{\link{polyfreqs}}.
}

