% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data_desc.R
\docType{data}
\name{ref_reads}
\alias{ref_reads}
\title{A dataset of 10 individuals sampled at 2 loci with reference read counts simulated from a binomial distribution with underlying allele frequency 0.4.}
\format{A 10 x 2 matrix}
\usage{
ref_reads
}
\description{
A dataset of 10 individuals sampled at 2 loci with reference read counts simulated from a binomial distribution with underlying allele frequency 0.4.
}
\keyword{datasets}

