% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simple_freqs.R
\name{simple_freqs}
\alias{simple_freqs}
\title{Point estimation of allele frequencies based on read counts}
\usage{
simple_freqs(tM, rM)
}
\arguments{
\item{tM}{Total reads matrix: matrix containing the total number of reads mapping to each locus for each individual.}

\item{rM}{Reference reads marix: matrix containing the number of reference reads mapping to each locus for each individual.}
}
\value{
A vector of allele frequencies, one for each locus. Named \code{allele_freqs_hat}.
}
\description{
\code{simple_freqs} estimates allele frequencies based on read count ratios.
}
\author{
Paul Blischak
}

