% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data_desc.R
\docType{data}
\name{total_reads}
\alias{total_reads}
\title{A dataset of 10 individuals sampled at 2 loci with 20 reads per individual per locus.}
\format{A 10 x 2 matrix.}
\usage{
total_reads
}
\description{
A dataset of 10 individuals sampled at 2 loci with 20 reads per individual per locus.
}
\keyword{datasets}

