% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/point_Hexp.R
\name{point_Hexp}
\alias{point_Hexp}
\title{Calculate expected heterozygosity from point estimates of allele frequencies}
\usage{
point_Hexp(freqs, ploidy)
}
\arguments{
\item{freqs}{A vector of estimated allele frequencies returned from the function \code{\link{simple_freqs}}.}

\item{ploidy}{The ploidy level of individuals in the population (must be >= 2).}
}
\value{
A list with two items: the overall expected heterozygosity across loci (\code{Hexp}) and a vector with the per locus estimates of expected heterozygosity (\code{per_locus_Hexp}).
}
\description{
Uses point estimates of allele frequencies calculated with \code{\link{simple_freqs}} to calculate the per locus and overall expected heterozygosity.
}

