% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_Hobs.R
\name{point_Hobs}
\alias{point_Hobs}
\title{Estimation of observed heterozygosity}
\usage{
point_Hobs(genotypes, ploidy)
}
\arguments{
\item{genotypes}{A matrix of estimated genotypes returned from the function \code{\link{get_map_genotypes}}.}

\item{ploidy}{The ploidy level of individuals in the population (must be >= 2).}
}
\value{
Returns per locus estimates of observed heterozygosity (\code{per_locus_Hobs}).
}
\description{
\emph{INTERNAL}: Estimates a posterior distribution for the per locus observed heterozygosity using the unbiased estimator of Hardy (2015) and the poterior samples of genotypes calculated by \code{\link{polyfreqs}}.
}
\details{
Posterior distributions for the per locus observed heterozygosity are automatically calculated and returned by the \code{\link{polyfreqs}} function.
}
\references{
Hardy, OJ. 2015. Population genetics of autopolyploids under a mixed mating model and the estimation of selfing rate. \emph{Molecular Ecology Resources}, doi: 10.1111/1755-0998.12431.
}

