% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_desc.R
\docType{data}
\name{ref_reads}
\alias{ref_reads}
\title{Reference reads matrix}
\format{A 10 x 2 matrix}
\usage{
data(ref_reads)
}
\description{
A dataset of 10 individuals sampled at 2 loci with
reference read counts simulated from a binomial distribution
(Eq. 1 in Blischak \emph{et al}.) with an underlying allele frequency of 0.4. Used for package testing.
}
\references{
Blischak PD, LS Kubatko and AD Wolfe. Accounting for genotype uncertainty in the estimation of allele frequencies in autopolyploids. \emph{In revision}.
}
\keyword{datasets}

