% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sim_reads.R
\name{sim_reads}
\alias{sim_reads}
\title{Simulation of sequencing read counts and genotypes}
\usage{
sim_reads(pVec, N_ind, coverage, ploidy, error)
}
\arguments{
\item{pVec}{A vector of allele frequencies strung together using the concatenate function.}

\item{N_ind}{The number of individuals to simulate}

\item{coverage}{The average number of sequenced simulated per individual per locus (Poisson distributed)}

\item{ploidy}{The ploidy level of individuals in the population.}

\item{error}{The probability of a sequencing error.}
}
\description{
Simulates read counts for under the model of Blischak et al. (2015).
}

