% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sim_reads.R
\name{sim_reads}
\alias{sim_reads}
\title{Simulation of sequencing read counts and genotypes}
\usage{
sim_reads(pVec, N_ind, coverage, ploidy, error)
}
\arguments{
\item{pVec}{A vector of allele frequencies strung together using the concatenate function.}

\item{N_ind}{The number of individuals to simulate.}

\item{coverage}{The average number of sequences simulated per individual per locus (Poisson distributed).}

\item{ploidy}{The ploidy level of individuals in the population.}

\item{error}{The level of sequencing error. A fixed constant.}
}
\value{
A list of 3 matrices: (1) \code{genos} - a matrix of the simulated genotypes, (2) \code{tot_read_mat} - a matrix of the simulated number of total reads, (3) \code{ref_read_mat} - a matrix of the simulated number of reference reads.
}
\description{
Simulates genotypes and read counts under the model of Blischak et al. (2015).
}
\references{
Blischak PD, Kubatko LS, Wolfe AD. 2015. Accounting for genotype uncertainty in the estimation of allele frequencies in autopolyploids. \emph{In review}. bioRxiv, \strong{doi}:####.
}

