% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_desc.R
\docType{data}
\name{total_reads}
\alias{total_reads}
\title{Total reads matrix}
\format{A 10 x 2 matrix.}
\usage{
data(total_reads)
}
\description{
A dataset of 10 individuals sampled at 2 loci with 20 reads per individual per locus. Used for package testing.
}
\keyword{datasets}

